add missing image, rename to brushing demo, add checkbox to enableBrushing, fix tooptip missing info
