+++ examples/brushing/arc-brushing-layer/index.js
