+++ examples/brushing/scatterplot-brushing-layer/index.js
