+++ examples/brushing/scatterplot-brushing-layer/scatterplot-brushing-layer-fragment.glsl.js
